Fix most of message and buffer sizes issues.
Buffer length was still on 16 bits while the message length was on 32. Some parameters of functions were still on 16 bits or on ints. Explicit the types wherever necessary.

+++ lib/include/mqtt_msg.h
@@ -81,7 +81,7 @@ typedef struct mqtt_connection
 
     uint16_t message_id;
     uint8_t* buffer;
-    uint16_t buffer_length;
+    uint32_t buffer_length;
 
 } mqtt_connection_t;
 
@@ -107,14 +107,14 @@ static inline int mqtt_get_dup(uint8_t* buffer) { return (buffer[0] & 0x08) >> 3
 static inline int mqtt_get_qos(uint8_t* buffer) { return (buffer[0] & 0x06) >> 1; }
 static inline int mqtt_get_retain(uint8_t* buffer) { return (buffer[0] & 0x01); }
 
-void mqtt_msg_init(mqtt_connection_t* connection, uint8_t* buffer, uint16_t buffer_length);
-uint32_t mqtt_get_total_length(uint8_t* buffer, uint16_t length);
+void mqtt_msg_init(mqtt_connection_t* connection, uint8_t* buffer, uint32_t buffer_length);
+uint32_t mqtt_get_total_length(uint8_t* buffer, uint32_t length);
 const char* mqtt_get_publish_topic(uint8_t* buffer, uint32_t* length);
 const char* mqtt_get_publish_data(uint8_t* buffer, uint32_t* length);
-uint16_t mqtt_get_id(uint8_t* buffer, uint16_t length);
+uint16_t mqtt_get_id(uint8_t* buffer, uint32_t length);
 
 mqtt_message_t* mqtt_msg_connect(mqtt_connection_t* connection, mqtt_connect_info_t* info);
-mqtt_message_t* mqtt_msg_publish(mqtt_connection_t* connection, const char* topic, const char* data, int data_length, int qos, int retain, uint16_t* message_id);
+mqtt_message_t* mqtt_msg_publish(mqtt_connection_t* connection, const char* topic, const char* data, uint32_t data_length, int qos, int retain, uint16_t* message_id);
 mqtt_message_t* mqtt_msg_puback(mqtt_connection_t* connection, uint16_t message_id);
 mqtt_message_t* mqtt_msg_pubrec(mqtt_connection_t* connection, uint16_t message_id);
 mqtt_message_t* mqtt_msg_pubrel(mqtt_connection_t* connection, uint16_t message_id);
